+++ tests/expected/setget.gd
@@ -3,7 +3,7 @@
 var a: float = 10.0:
 	set(value):
 		a = value
-	get:
+	get():
 		return a
 var second: int = 10