+++ tests/input/match_case.gd
@@ -27,3 +27,6 @@ func foo():
 		Type.LEFT: animation.play("left")
 		Type.RIGHT: animation.play("right")
 		_: animation.play("right")
+	match value:
+		1 or 2:
+			print("1 or 2")